Admin side purchases complete

+++ server/app.py
@@ -127,13 +127,6 @@ class StoreById(Resource):
       else:
          return {"message" : "Not Authorized"}, 401
 
-
-# class StoresByOwner(Resource):
-#    def get(self, userid):
-#       stores = Store.query.filter_by(user_id = userid).all()
-#       stores_dict = [store.to_dict() for store in stores]
-#       return stores_dict, 200
-
    
 class Items(Resource):
    def get(self):
@@ -176,10 +169,14 @@ class ItembyId(Resource):
          return {}, 404
       
    def delete(self, itemid):
+      user = User.query.filter(User.id == session.get('user_id')).first()
       item = Item.query.filter_by(id=itemid).first()
-      db.session.delete(item)
-      db.session.commit()
-      return {}, 204
+      if user.isAdmin and item.store in user.stores:
+         db.session.delete(item)
+         db.session.commit()
+         return {}, 204
+      else: 
+         return {"message" : "Not Authorized"}, 401
    
    def patch(self, itemid):
       user = User.query.filter(User.id == session.get('user_id')).first()
@@ -214,22 +211,21 @@ class Purchases(Resource):
       return newPurchase.to_dict(), 201
 
 
-   
 class PurchasesByUser(Resource):
    def get(self, userid):
       purchases = Purchase.query.filter_by(user_id = userid).all()
       purchases_dict = [purchase.to_dict(rules=('-user',)) for purchase in purchases]
       return purchases_dict, 200
    
-class PurchasesByStore(Resource):
-   def get(self, storeid):
-      store = Store.query.filter_by(id = storeid).first()
-      items = store.items
-      purchases = []
-      for item in items:
-         purchases.extend(Purchase.query.filter_by(item_id = item.id).all())
-      purchases_dict = [purchase.to_dict() for purchase in purchases]
-      return purchases_dict, 200
+# class PurchasesByStore(Resource):
+#    def get(self, storeid):
+#       store = Store.query.filter_by(id = storeid).first()
+#       items = store.items
+#       purchases = []
+#       for item in items:
+#          purchases.extend(Purchase.query.filter_by(item_id = item.id).all())
+#       purchases_dict = [purchase.to_dict() for purchase in purchases]
+#       return purchases_dict, 200
    
 api.add_resource(Signup, '/api/signup')
 api.add_resource(CheckSession, '/api/check_session')
@@ -237,10 +233,9 @@ api.add_resource(Login, '/api/login')
 api.add_resource(Logout, '/api/logout')
 api.add_resource(StoreList, '/api/stores')
 api.add_resource(StoreById, '/api/stores/<int:storeid>')
-# api.add_resource(StoresByOwner, '/api/stores/owner/<int:userid>')
 api.add_resource(Items, '/api/items')
 api.add_resource(ItembyId, '/api/items/<int:itemid>')
-api.add_resource(PurchasesByStore, '/api/purchases/stores/<int:storeid>')
+# api.add_resource(PurchasesByStore, '/api/purchases/stores/<int:storeid>')
 api.add_resource(Purchases, '/api/purchases')
 api.add_resource(PurchasesByUser, '/api/purchases/<int:userid>')
